sample.spring.boot.parent/sample.springboot.properties/src/main/java/com/ay/sample/springboot/properties/PropertiesExpress.java
package com.ay.sample.springboot.properties;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Component;

@Component
public class PropertiesExpress {
    @Value("${sample.conf.title}")
    private String title;

    @Value("${sample.conf.type}")
    private String type;

    @Value("${sample.conf.login[name]}")
    private String userName;

    public String getUserName() {
        return userName;
    }

    public void setUserName(String userName) {
        this.userName = userName;
    }

    public String getTitle() {
        return title;
    }

    public void setTitle(String title) {
        this.title = title;
    }

    public String getType() {
        return type;
    }

    public void setType(String type) {
        this.type = type;
    }
}
